+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @solana/example-nextjs
 
+## 0.0.1-rc.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @solana/client@1.0.0-rc.1
+  - @solana/react-hooks@1.0.0-rc.1
+
 ## 0.0.1-rc.0
 
 ### Patch Changes